requirements.txt
Flask
pytest
selenium
requests
google-generativeai
psutil
werkzeug
pytest-cov
coverage